AT-9683: updating to use emit
